refactor(uart): align model layout and documentation
