+++ driver/ch/ch32_uart.hpp
@@ -6,9 +6,9 @@
 #include "ch32_uart_def.hpp"
 #include "libxr_def.hpp"
 #include "libxr_rw.hpp"
-#include "uart.hpp"
-#include "uart_circular_dma_rx_model.hpp"
-#include "uart_dma_tx_model.hpp"
+#include "uart/uart.hpp"
+#include "uart/uart_circular_dma_rx_model.hpp"
+#include "uart/uart_dma_tx_model.hpp"
 
 namespace LibXR
 {
@@ -55,9 +55,10 @@ class CH32UART : public UART
 
  private:
   /**
-   * @brief Program and start the CH32 circular UART RX DMA channel.
-   * @param data DMA-writable receive buffer.
-   * @param size Receive buffer capacity in bytes.
+   * @brief 配置并启动 CH32 UART 循环 RX DMA 通道 / Configure and start the CH32 UART
+   * circular RX DMA channel
+   * @param data DMA 可写的接收缓冲区 / DMA-writable receive buffer
+   * @param size 接收缓冲区字节数 / Receive buffer capacity in bytes
    */
   void StartCircularDmaRx(uint8_t* data, size_t size)
   {
@@ -67,24 +68,28 @@ class CH32UART : public UART
   }
 
   /**
-   * @brief Return the CH32 RX DMA remaining transfer count.
+   * @brief 获取 CH32 RX DMA 剩余传输计数 / Get the CH32 RX DMA remaining count
+   * @return DMA 尚未写入的字节数 / Number of bytes not yet written by DMA
    */
   [[nodiscard]] size_t GetCircularDmaRxRemaining() const { return dma_rx_channel_->CNTR; }
 
   /**
-   * @brief Prepare CH32 circular DMA RX storage for CPU access.
-   * @param data DMA receive buffer start address.
-   * @param size Receive buffer capacity in bytes.
-   * @note CH32 requires no cache maintenance for this buffer.
+   * @brief 为 CPU 访问准备 CH32 循环 DMA RX 存储区 / Prepare CH32 circular DMA RX storage
+   * for CPU access
+   * @param data DMA 接收缓冲区起始地址 / DMA receive buffer start address
+   * @param size 接收缓冲区字节数 / Receive buffer capacity in bytes
+   * @note 该 CH32 缓冲区不需要缓存维护 / This CH32 buffer requires no cache maintenance
    */
   void PrepareCircularDmaRxForCpu(uint8_t*, size_t) {}
 
   /**
-   * @brief Program and start one active UART TX DMA payload.
-   * @param data DMA-readable payload buffer.
-   * @param size Payload size in bytes.
-   * @param block Active double-buffer block index; unused by CH32 DMA.
-   * @return True after the DMA channel is enabled.
+   * @brief 配置并启动一个 active UART TX DMA 载荷 / Configure and start one active UART
+   * TX DMA payload
+   * @param data DMA 可读的载荷缓冲区 / DMA-readable payload buffer
+   * @param size 载荷字节数 / Payload size in bytes
+   * @param block active 双缓冲块索引，CH32 DMA 不使用 / Active double-buffer block index,
+   * unused by CH32 DMA
+   * @return DMA 通道启用后返回 true / True after the DMA channel is enabled
    */
   bool StartDmaTx(uint8_t* data, size_t size, int block);
 };
