+++ src/driver/usb/device/cdc/cdc_base.hpp
@@ -3,7 +3,7 @@
 
 #include "dev_core.hpp"
 #include "libxr_def.hpp"
-#include "uart.hpp"
+#include "uart/uart.hpp"
 #include "usb/core/desc_cfg.hpp"
 
 namespace LibXR::USB
