+++ src/driver/usb/device/cdc/cdc_to_uart.hpp
@@ -3,7 +3,7 @@
 #include "cdc_uart.hpp"
 #include "libxr_cb.hpp"
 #include "libxr_rw.hpp"
-#include "uart.hpp"
+#include "uart/uart.hpp"
 
 namespace LibXR::USB
 {
