+++ test/framework/test_main_sets.hpp
@@ -64,10 +64,6 @@ inline constexpr GroupedTestCase kMainTestCases[] = {
     {"core_tests", {"callback", &RunVoidEntry<test_cb>, false}},
     {"core_tests", {"pipe", &RunVoidEntry<test_pipe>, false}},
     {"core_tests", {"rw", &RunVoidEntry<test_rw>, false}},
-    {"core_tests",
-     {"uart_circular_dma_rx_model", &RunVoidEntry<test_uart_circular_dma_rx_model>,
-      false}},
-    {"core_tests", {"uart_dma_tx_model", &RunVoidEntry<test_uart_dma_tx_model>, false}},
     {"core_tests", {"memory", &RunVoidEntry<test_memory>, false}},
     {"core_tests", {"color", &RunVoidEntry<test_color>, false}},
     {"core_tests", {"time", &RunVoidEntry<test_time>, false}},
